+++ wrappers/mbed/hw/nucleo_wl55jc.h
@@ -81,8 +81,13 @@ namespace LDL {
 
             public:
 
+                /** Create
+                 *
+                 * @param[in] tx_gain       antenna gain (dB x 100)
+                 *
+                 * */
                 NucleoWL55JC(
-                    int16_t tx_gain = 0
+                    int16_t tx_gain = 200
                 )
                     :
                     Radio(),