+++ wrappers/mbed/hw/sx1272mb2xas.h
@@ -46,8 +46,13 @@ namespace LDL {
 
             public:
 
+                /** Create
+                 *
+                 * @param[in] tx_gain       antenna gain (dB x 100)
+                 *
+                 * */
                 SX1272MB2XAS(
-                    int16_t tx_gain = 0
+                    int16_t tx_gain = 200
                 )
                     :
                     Radio(),