+++ wrappers/mbed/sx1261.h
@@ -42,7 +42,7 @@ namespace LDL {
              * @param[in] dio1
              * @param[in] dio2
              * @param[in] dio3
-             * @param[in] tx_gain
+             * @param[in] tx_gain       antenna gain (dB x 100)
              * @param[in] regulator
              * @param[in] voltage
              * @param[in] xtal
@@ -57,7 +57,7 @@ namespace LDL {
                 PinName dio1,
                 PinName dio2,
                 PinName dio3,
-                int16_t tx_gain = 0,
+                int16_t tx_gain = 200,
                 enum ldl_sx126x_regulator regulator = LDL_SX126X_REGULATOR_LDO,
                 enum ldl_sx126x_txen txen = LDL_SX126X_TXEN_ENABLED,
                 enum ldl_sx126x_voltage voltage = LDL_SX126X_VOLTAGE_1V6,